+++ cmd/starter/c/starter.c
@@ -3,7 +3,7 @@
     Apptainer a Series of LF Projects LLC.
     For website terms of use, trademark policy, privacy policy and other
     project policies see https://lfprojects.org/policies
-  Copyright (c) 2018-2020, Sylabs, Inc. All rights reserved.
+  Copyright (c) 2018-2023, Sylabs, Inc. All rights reserved.
 
   This software is licensed under a 3-clause BSD license.  Please
   consult LICENSE.md file distributed with the sources of this project regarding
@@ -316,7 +316,9 @@ static void apply_privileges(struct privileges *privileges, struct capabilities 
                 fatalf("Failed to set GID %d: %s\n", targetGID, strerror(errno));
             }
 
-            if ( privileges->numGID > 1 ) {
+            if ( privileges->noSetgroups ) {
+                debugf("Skipping setgroups due to configuration.\n");
+            } else {
                 debugf("Set %d additional group IDs\n", privileges->numGID);
                 if ( setgroups(privileges->numGID, privileges->targetGID) < 0 ) {
                     fatalf("Failed to set additional groups: %s\n", strerror(errno));