+++ internal/pkg/runtime/engine/fakeroot/config/config.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2019, Sylabs Inc. All rights reserved.
+// Copyright (c) 2019-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -15,8 +15,9 @@ const Name = "fakeroot"
 // EngineConfig is the config for the fakeroot engine used to execute
 // a command in a fakeroot context
 type EngineConfig struct {
-	Args     []string `json:"args"`
-	Envs     []string `json:"envs"`
-	Home     string   `json:"home"`
-	BuildEnv bool     `json:"buildEnv"`
+	Args        []string `json:"args"`
+	Envs        []string `json:"envs"`
+	Home        string   `json:"home"`
+	BuildEnv    bool     `json:"buildEnv"`
+	NoSetgroups bool     `json:"NoSetgroups"`
 }