+++ internal/pkg/runtime/launch/options.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2022-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -120,6 +120,8 @@ type launchOptions struct {
 
 	// Fakeroot enables the fake root mode, using user namespaces and subuid / subgid mapping.
 	Fakeroot bool
+	// NoSetgroups disables calling setgroups for the fakeroot user namespace.
+	NoSetgroups bool
 	// Boot enables execution of /sbin/init on startup of an instance container.
 	Boot bool
 	// NoInit disables shim process when PID namespace is used.
@@ -435,6 +437,14 @@ func OptCwdPath(p string) Option {
 func OptFakeroot(b bool) Option {
 	return func(lo *launchOptions) error {
 		lo.Fakeroot = b
+		return nil
+	}
+}
+
+// OptNoSetgroups disables calling setgroups for the fakeroot user namespace.
+func OptNoSetgroups(b bool) Option {
+	return func(lo *launchOptions) error {
+		lo.NoSetgroups = b
 		return nil
 	}
 }